NEW: Posterization Effect
Posterization in this case works only with LUMINANCE, which means, it only cut VALUE of color (Not Hue or Saturation).

A bit of explanation for ones who are reading all these: we take RGB in the end of all changes of our shader, convert it to HSV (Hue, Saturation, Value), and cut Value (or Brightness) into slices and asign them with one specific color each. And then convert it back to RGB, but this time Brightness can only be in steps.

For some games, same workflow must be applied to HUE and Saturation as well.

Additionally, I'm using predefined values for slices, but you can turn them into parameters.