Panini Projection and Various Fixes
Meet Panini Projection!

It is used with wide angle cameras to compensate the distance to the center of the image by making it closer. For example, I'm going to use a 28mm (in physical units in Godot) lens with 0.4 scale of my Panini slider.

My implementation can't overscan the image in higher resolution, so with too high values it will lose details noticeably. But with proper setup it allows to use fairly wide angles with acceptable sharpness. It will help to avoid motion sickness in first person games.

Chromatic Aberrations now have falloff in the center (in addition to the one that was lowering the effect for dark areas). Both sliders are 0.9 by default, which allows you to use very high aberrations without motion sickness.

WARNING: Bloom Booster has been disabled in this commit, it will be created from scratch tomorrow (some blocks are prepared for it). We need a better solution (along with anamorphic lens flares).